requirements.txt
pandas>=2.0
numpy>=1.24
pyyaml>=6.0
requests>=2.31
python-dotenv>=1.0
geopandas>=0.14
shapely>=2.0
scikit-learn>=1.3
hdbscan>=0.8.33
tqdm>=4.66
pyarrow>=14.0
folium>=0.15
branca>=0.7
beautifulsoup4>=4.12
lxml>=4.9
snowflake-connector-python>=3.6
